pyproject.toml
[tool.poetry]
name = "corp-staff"
version = "0.1.0"
description = "company employers API"
authors = ["Maksim Semenyuk <semenyuk.mv85@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.3"
pillow = "^9.5.0"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
